+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
 ## [In Development](https://github.com/ppfeufer/eve-online-intel-tool/tree/development)
 [Full Changelog](https://github.com/ppfeufer/eve-online-intel-tool/compare/v1.4.1...development)
-- in development
+### Added
+- Getting alliance logos, corporation logos and pilot portraits from ESI now
 
 ## [v1.4.1](https://github.com/ppfeufer/eve-online-intel-tool/releases/tag/v1.4.1) - 2018-11-17
 [Full Changelog](https://github.com/ppfeufer/eve-online-intel-tool/compare/v1.4.0...v1.4.1)